Fix FormData PR CI drift
