+++ crates/perry-runtime/src/object/mod.rs
@@ -36,6 +36,7 @@ mod descriptors;
 mod field_get_set;
 mod field_set_by_name;
 mod global_this;
+mod global_this_tables;
 mod groupby;
 pub(crate) mod has_own_helpers;
 mod instanceof;
@@ -67,6 +68,7 @@ pub use descriptors::*;
 pub use field_get_set::*;
 pub use field_set_by_name::*;
 pub use global_this::*;
+pub(crate) use global_this_tables::*;
 pub use groupby::*;
 pub use instanceof::*;
 pub use native_call_method::*;
